requirements.txt
ultralytics>=8.0.0
streamlit>=1.30.0
pillow
numpy
opencv-python-headless
